refactor stores
