--- tournament-desktop/src/jvmMain/kotlin/info/marozzo/tournament/desktop/application/store/TournamentIntent.kt
@@ -1,3 +1,0 @@
-package info.marozzo.tournament.desktop.application.store
-
-internal sealed interface TournamentIntent
