+++ tournament-desktop/src/jvmMain/kotlin/info/marozzo/tournament/desktop/application/stores/tournament/TournamentExecutor.kt
@@ -1,4 +1,4 @@
-package info.marozzo.tournament.desktop.application.store
+package info.marozzo.tournament.desktop.application.stores.tournament
 
 import com.arkivanov.mvikotlin.extensions.coroutines.CoroutineExecutor
