+++ tournament-desktop/src/jvmMain/kotlin/info/marozzo/tournament/desktop/application/stores/tournament/TournamentIntent.kt
@@ -1,0 +1,5 @@
+package info.marozzo.tournament.desktop.application.stores.tournament
+
+internal sealed interface TournamentIntent
+
+data class CreateNewEvent(val name: String): TournamentIntent
