+++ tournament-desktop/src/jvmMain/kotlin/info/marozzo/tournament/desktop/application/stores/tournament/TournamentMessage.kt
@@ -1,0 +1,3 @@
+package info.marozzo.tournament.desktop.application.stores.tournament
+
+internal sealed interface TournamentMessage
